파일 이어받기 구현1(#8)

+++ TCP_CLASS.py
@@ -74,22 +74,37 @@ class TcpSocket :
     def send_file(self, file_name) :
 
         hasher = hashlib.sha224()
-        file_size = os.path.getsize(file_name)
 
-        length = int((file_size) / (self.BUFSIZE) + 0.5) + 1
-        self.send_message(str(length))
+        reply = self.receive_message() # 파일 존재 여부에 대한 메시지 받기 ( None : 없음, Exist : 존재 )
+        file = open(file_name, "rb")
+        if reply == "Exist" :
+
+            jump_size = self.receive_message() # 클라이언트에 존재하던 파일의 크기를 받음
+            my_hash_value = self.get_hash_value(file_name, jump_size)
+            self.send_message(str(my_hash_value)) # 무결성 체크를 위해 클라이언트에게 해쉬값 전송
+            file.read(jump_size) # jump_size 만큼 파일 포인터 점프
+
+            file_size = os.path.getsize(file_name)
+            length = int(((file_size - jump_size)) / (self.BUFSIZE) + 0.5) + 1
+            self.send_message(str(length))
+
+        else :
+
+            file_size = os.path.getsize(file_name)
+            length = int((file_size) / (self.BUFSIZE) + 0.5) + 1
+            self.send_message(str(length))
 
         try:
-            print file_name.split('/')[-1].decode('cp949').encode('utf-8'), ; print '을 전송합니다.'
+            print file_name.split('/')[-1].decode('cp949').encode('utf-8'),; print '을 전송합니다.'
         except:
-            print file_name.split('/')[-1], ; print '을 전송합니다.'
+            print file_name.split('/')[-1],; print '을 전송합니다.'
 
-        with open(file_name, "rb") as f :
-            buf = f.read(self.BUFSIZE)
+        while file :
+            buf = file.read(self.BUFSIZE)
             while buf :
                 hasher.update(buf)
                 self.send_message(buf)
-                buf = f.read(self.BUFSIZE)
+                buf = file.read(self.BUFSIZE)
 
         self.send_message(str(file_size))
         self.send_message(str(hasher.hexdigest()))
@@ -155,19 +170,65 @@ class TcpSocket :
                 break
 
 
+    def get_hash_value(self, file_name, file_size = None) :
+
+        if file_size is None :
+            file_size = os.path.getsize(file_name)
+
+        try :
+            hasher = hashlib.sha224()
+
+            with open(file_name, "rb") as file:
+                for i in range(0, file_size, self.BUFSIZE) :
+                    buf = file.read(self.BUFSIZE)
+                    if buf : hasher.update(buf)
+            return hasher.hexdigest()
+        except Exception, e :
+            print e
+            sys.exit(-1)
 
     def receive_file(self, file_name, directory = None) :
 
         start = datetime.datetime.now()
-        out_file = open(file_name, "wb")
-        length = int(self.receive_message())
 
-        for i in range(length) :
-            data = self.receive_message()
-            while not data :
+        if os.path.exists(file_name) is True :
+
+            self.send_message("Exist") # 파일 존재 여부 확인 메시지 보내기
+
+            out_file = open(file_name, 'a')
+
+            file_size = os.path.getsize(file_name)
+            self.send_message(str(file_size)) # 서버로 내가 가지고있던 파일 크기를 보냄 ( 서버에서는 size 만큼 파일 포인터 점프 )
+
+            ## 존재하던 파일의 무결성 체크 ##
+            my_hash_value = self.get_hash_value(file_name, file_size)
+            recv_hash_value = self.receive_message()
+
+            if str(my_hash_value) != str(recv_hash_value) :
+                print "존재하던 파일이 손상 되었습니다. 삭제한 후에 다시 시도해주세요!"
+                sys.exit(0)
+
+
+            length = int(self.receive_message())
+            for i in range(length) :
                 data = self.receive_message()
-            out_file.write(data)
-        out_file.close()
+                while not data :
+                    data = self.receive_message()
+                out_file.write(data)
+            out_file.close()
+
+        else :
+            self.send_message("None")  # 파일 존재 여부 확인 메시지 보내기
+            out_file = open(file_name, "wb")
+            length = int(self.receive_message())
+
+            for i in range(length) :
+                data = self.receive_message()
+                while not data :
+                    data = self.receive_message()
+                out_file.write(data)
+            out_file.close()
+
         end = datetime.datetime.now()
 
         original_file_size = self.receive_message()
